requirements.txt
Django==2.0.6
djangorestframework==3.8.2
django-flat-responsive==2.0
django-cors-headers==2.3.0
django-import-export==1.0.1